Constants and adding application as parameter
